feat: Support safepoint poll for AArch64 (#58)
## Related issue(s):

issue #55  #63

## What this PR does / why we need it:

Add safepoint poll support for AArch64. Also fixes #63, making the VM
call relocation type `runtime_call_type`.

For aarch64, to set the `last_Java_pc` correctly, we have to preserve
the return address of the Call VM stub [1].
The current implementation of generate_call_VM only saves
`last_Java_sp`, and for LLVM, it's not easy to obtain the return address
after the call in the stub. So I added the statepoint attribution for
the call in the Call VM stub, to get extra space for `last_Java_pc`
saving logic.

For x86, we can simply set the call site size of `VM_CALL` to 0 to avoid
patching.


1.
https://github.com/jeandle/jeandle-jdk/blob/3e861739f63280b3e843714b615b9f46880b0a10/src/hotspot/cpu/aarch64/aarch64.ad#L3776-L3784

Testing:
- [x] debug/release build and `tier1_compiler_jeandle` on aarch64
- [x] debug build and `tier1_compiler_jeandle` on x64

---------

Signed-off-by: Feilong Jiang <jiangfeilong@huawei.com>